dba33d: #i102563# use rootURL to resolve the given URI and some refactoring

+++ reportbuilder/java/com/sun/star/report/pentaho/output/text/PageContext.java
@@ -75,7 +75,7 @@ public class PageContext
             // TODO: IS this code correct? Why not columnCount = pc.getColumnCount(); ?
             if (columnCount != null)
             {
-                return columnCount.intValue();
+                return columnCount;
             }
             pc = pc.getParent();
         }
